# apps/backend/.openapi.after.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Synthetic Data Platform",
    "description": "Synthetic Data Platform API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/projects/{project_id}/infer/providers": {
      "post": {
        "tags": [
          "infer"
        ],
        "summary": "Infer Providers For Project",
        "operationId": "infer_providers_for_project_api_v1_projects__project_id__infer_providers_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-User-Sub",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Sub"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InferProvidersRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InferProvidersResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/infer/providers": {
      "post": {
        "tags": [
          "infer"
        ],
        "summary": "Infer Providers Nonscoped",
        "operationId": "infer_providers_nonscoped_api_v1_infer_providers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InferProvidersRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InferProvidersResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ColumnProviderSuggestions": {
        "properties": {
          "table": {
            "type": "string",
            "title": "Table"
          },
          "column": {
            "type": "string",
            "title": "Column"
          },
          "suggestions": {
            "items": {
              "$ref": "#/components/schemas/ProviderSuggestion"
            },
            "type": "array",
            "title": "Suggestions"
          }
        },
        "type": "object",
        "required": [
          "table",
          "column"
        ],
        "title": "ColumnProviderSuggestions"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InferColumn": {
        "properties": {
          "table": {
            "type": "string",
            "title": "Table",
            "description": "Table name"
          },
          "column": {
            "type": "string",
            "title": "Column",
            "description": "Column name"
          },
          "dtype": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dtype",
            "description": "Normalized data type"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Free-text description"
          }
        },
        "type": "object",
        "required": [
          "table",
          "column"
        ],
        "title": "InferColumn"
      },
      "InferProvidersRequest": {
        "properties": {
          "columns": {
            "items": {
              "$ref": "#/components/schemas/InferColumn"
            },
            "type": "array",
            "title": "Columns"
          },
          "llm": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LLMRequest"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "InferProvidersRequest"
      },
      "InferProvidersResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/ColumnProviderSuggestions"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "title": "InferProvidersResponse"
      },
...
      "LLMRequest": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": false
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider",
            "description": "openai|anthropic|ollama|lmstudio"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "temperature": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature"
          }
        },
        "type": "object",
        "title": "LLMRequest"
      },
...
      "ProviderSuggestion": {
        "properties": {
          "provider_config": {
            "title": "Provider Config"
          },
          "score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Score"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "description": "'heuristic' or 'LLM'"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "reasons": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasons"
          },
          "pii": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pii"
          }
        },
        "type": "object",
        "required": [
          "provider_config",
          "score",
          "source"
        ],
        "title": "ProviderSuggestion"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}